+++ GEECS-Core/geecs_core/db/__init__.py
@@ -2,5 +2,12 @@
 
 from geecs_core.db.geecs_db import GeecsDb
 from geecs_core.db.scalar_policy import GeecsDbScalarPolicy, ScalarPolicyProvider
+from geecs_core.db.settables import NumericSettable, numeric_settables
 
-__all__ = ["GeecsDb", "GeecsDbScalarPolicy", "ScalarPolicyProvider"]
+__all__ = [
+    "GeecsDb",
+    "GeecsDbScalarPolicy",
+    "NumericSettable",
+    "ScalarPolicyProvider",
+    "numeric_settables",
+]